+++ Weathy/Weathy/Sources/Models/Calendar/MonthlyWeathyData.swift
@@ -11,17 +11,3 @@ struct MonthlyWeathyData: Codable {
     var calendarOverviewList: [CalendarOverview?]?
     var message: String
 }
-
-// MARK: - CalendarOverviewList
-struct CalendarOverview: Codable {
-    var id: Int
-    var date: String
-    var stampId: Int
-    var temperature: CalendarTemperature
-
-}
-
-// MARK: - Temperature
-struct CalendarTemperature: Codable {
-    var maxTemp, minTemp: Int
-}
