+++ Weathy/Weathy/Sources/VCs/ModifyWeathy/ModifyWeathyTagVC.swift
@@ -11,7 +11,7 @@ class ModifyWeathyTagVC: UIViewController {
     
     //MARK: - Custom Variables
     
-    var weathyData: CalendarWeathy?
+    var weathyData: Weathy?
     var dateString: String = "0000-00-00"
     var locationCode: CLong = 1141000000
